bookstore-micro05/auth-service/requirements.txt
fastapi
uvicorn[standard]
httpx
PyJWT
